requirements.txt
google-cloud-storage
pymongo
python-dotenv
requests
pillow